+++ src/Inviqa/OneStock/Factory/ApiOperationFactory.php
@@ -7,7 +7,7 @@ use Inviqa\OneStock\Config;
 use Inviqa\OneStock\Order\OrderExporter;
 use Inviqa\OneStock\Order\OrderSanitizer;
 use Inviqa\OneStock\Order\Request\JsonRequestBuilder;
-use Inviqa\OneStock\RequestAgent;
+use Inviqa\OneStock\Agent\RequestAgent;
 
 class ApiOperationFactory
 {
